Se corrige el tipo de dato de cada numero, para el caso de obtener un promedio no entero

+++ Positive_Power/Program.cs
@@ -1,13 +1,13 @@
 ﻿/*  Pide al usuario cuatro números y muestra el promedio.   */
 
 Console.WriteLine("Digita el primer numero: ");
-int num1 = int.Parse(Console.ReadLine());
+double num1 = double.Parse(Console.ReadLine());
 Console.WriteLine("Digita el primer numero: ");
-int num2 = int.Parse(Console.ReadLine());
+double num2 = double.Parse(Console.ReadLine());
 Console.WriteLine("Digita el primer numero: ");
-int num3 = int.Parse(Console.ReadLine());
+double num3 = double.Parse(Console.ReadLine());
 Console.WriteLine("Digita el primer numero: ");
-int num4 = int.Parse(Console.ReadLine());
+double num4 = double.Parse(Console.ReadLine());
 
 
 Console.WriteLine($"El promedio de los 4 numeros digitados es: {(num1+num2+num3+num4)/4}");
